requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pymupdf
python-docx
google-generativeai
mysql-connector-python
python-dotenv
